+++ testing/rspec/spec/unit/shared/containing_models_unit_specs.rb
@@ -22,6 +22,108 @@ shared_examples_for 'a containing model' do
     expect(model.children).to_not include(nil)
   end
 
+
+  describe 'tree traversal' do
+
+    it 'is enumerable' do
+      expect(model).to be_a(Enumerable)
+      expect(model).to respond_to(:each)
+    end
+
+    it 'provides an Enumerator' do
+      expect(model.each).to be_a(Enumerator)
+    end
+
+
+    context 'with children' do
+
+
+      let(:root_model) do
+        m = clazz.new
+        m.extend(CukeModeler::ModelFactory)
+
+        # Not all models have a name, so making sure that this one does
+        def m.name
+          'top level model'
+        end
+
+        # Redefining what children are returned for the models rather than fiddling with all of the different kinds of
+        # child models that various model classes have
+        def m.children
+          childless_child_object = generate_model(mixins: [CukeModeler::Named],
+                                                  attributes: { name: 'child object 1' })
+
+          child_object_with_children = generate_model(mixins: [CukeModeler::Named, CukeModeler::ModelFactory],
+                                                      attributes: { name: 'child object 2' })
+
+          def child_object_with_children.children
+            nested_child_object = generate_model(mixins: [CukeModeler::Named],
+                                                 attributes: { name: 'grandchild object' })
+            [nested_child_object]
+          end
+
+          [childless_child_object, child_object_with_children]
+        end
+
+        m
+      end
+
+      it 'iterates through all of the models in its tree' do
+        names = []
+
+        root_model.each do |model_node|
+          names << model_node.name
+        end
+
+        expect(names).to match_array(['top level model', 'child object 1', 'child object 2', 'grandchild object'])
+      end
+
+      it 'iterates over the root model first' do
+        names = []
+
+        root_model.each do |model_node|
+          names << model_node.name
+        end
+
+        expect(names.first).to eq('top level model')
+      end
+
+    end
+
+    context 'without children' do
+
+      let(:parent_model) do
+        m = clazz.new
+
+        # Not all models have a name, so making sure that this one does
+        def m.name
+          'top level model'
+        end
+
+        # Redefining what children are returned for the model rather than fiddling with all of the different kinds of
+        # child models that various model classes have
+        def m.children
+          []
+        end
+
+        m
+      end
+
+      it 'iterates through all of the models in its tree' do
+        names = []
+
+        parent_model.each do |model_node|
+          names << model_node.name
+        end
+
+        expect(names).to match_array(['top level model'])
+      end
+
+    end
+
+  end
+
+
   it 'can execute code on all models in its model tree' do
     expect(model).to respond_to(:each_model)
   end